+++ underworld3/systems/poisson.pyx
@@ -36,7 +36,7 @@ class Poisson:
         # create private variables
         options = PETSc.Options()
         options.setValue("uprivate_petscspace_degree", degree) # for private variables
-        self.petsc_fe_u = PETSc.FE().createDefault(self.dm.getDimension(), 1, mesh.isSimplex, degree, "uprivate_", PETSc.COMM_WORLD)
+        self.petsc_fe_u = PETSc.FE().createDefault(mesh.dim, 1, mesh.isSimplex, degree, "uprivate_", PETSc.COMM_WORLD)
         self.petsc_fe_u_id = self.dm.getNumFields()
         self.dm.setField( self.petsc_fe_u_id, self.petsc_fe_u )